BE/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
sqlalchemy>=2.0.30
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
pydantic[email]>=2.7.4
pydantic-settings>=2.3.4
google-auth>=2.30.0
pyjwt>=2.8.0
python-multipart>=0.0.9
